catch case with no files from previous analyses

+++ R/clean.R
@@ -16,28 +16,37 @@ clean <- function(work.dir = getwd()) {
   models <- list.files(path = work.dir, pattern = "*.m_models.txt", full.names = TRUE)
   files <- list.files(path = work.dir, pattern = "*.mbmdr", full.names = TRUE)
   results <- list.files(path = work.dir, pattern = "*.result", full.names = TRUE)
+  topfile <- list.files(path = work.dir, pattern = "*.topfile", full.names = TRUE)
   registries <- grep("registries", list.dirs(path = work.dir, recursive = FALSE), value = TRUE)
+  topfiles <- grep("topfiles", list.dirs(path = work.dir, recursive = FALSE), value = TRUE)
+  permutations <- grep("permutation", list.dirs(path = work.dir, recursive = FALSE), value = TRUE)
 
-  message("Found the following files and directories from previous MB-MDR analyses:")
-  message(paste(oldstuff <- c(logs, models, files, results, registries), collapse = "\n"))
+  oldstuff <- c(logs, models, files, results, topfile, registries, topfiles, permutations)
 
-  answer <- readline("Delete [a]ll/[n]one/[s]ingle? ")
+  if(length(oldstuff) > 0) {
+    message("Found the following files and directories from previous MB-MDR analyses:")
+    message(paste(oldstuff, collapse = "\n"))
 
-  deleteSingle <- function(os) {
-    sapply(os, function(x) {
-      ans <- readline(paste("Delete", x, "? [y]es/[n]o/[c]ancel: "))
-      switch (ans,
-              y = invisible(unlink(x, recursive = TRUE, force = TRUE)),
-              n = return(),
-              c = stop("Cancelled!")
-      )
-    })
+    answer <- readline("Delete [a]ll/[n]one/[s]ingle? ")
+
+    deleteSingle <- function(os) {
+      sapply(os, function(x) {
+        ans <- readline(paste("Delete", x, "? [y]es/[n]o/[c]ancel: "))
+        switch (ans,
+                y = invisible(unlink(x, recursive = TRUE, force = TRUE)),
+                n = return(),
+                c = stop("Cancelled!")
+        )
+      })
+    }
+
+    switch (answer,
+            a = invisible(unlink(oldstuff, recursive = TRUE, force = TRUE)),
+            n = stop("Nothing done!"),
+            s = deleteSingle(oldstuff)
+    )
+  } else {
+    message("No files and directories from previous MB-MDR analyses found.")
   }
-
-  switch (answer,
-    a = invisible(unlink(oldstuff, recursive = TRUE, force = TRUE)),
-    n = stop("Nothing done!"),
-    s = deleteSingle(oldstuff)
-  )
 
 }
